Add RootResult object to store and show results of root finding
