+++ cxroots/RootFinder.py
@@ -10,7 +10,7 @@ import warnings
 
 from .IterativeMethods import iterateToRoot
 from .CountRoots import prod, RootError
-
+from .RootResult import RootResult
 
 def findRootsGen(originalContour, f, df=None, guessRoot=[], guessRootSymmetry=None, 
 	newtonStepTol=1e-14, newtonMaxIter=50, rootErrTol=1e-10,
@@ -315,97 +315,4 @@ def findRoots(originalContour, f, df=None, **kwargs):
 	rootFinder = findRootsGen(originalContour, f, df, **kwargs)
 	for roots, multiplicities, boxes, numberOfRemainingRoots in rootFinder:
 		pass
-	return roots, multiplicities
-
-def demo_findRoots(originalContour, f, df=None, automaticAnimation=False, saveFile=None, **kwargs):
-	"""
-	An interactive demonstration of the processess used to find all the roots
-	of a given function f within a given originalContour.
-	Shares key word arguments with :func:`cxroots.RootFinder.findRootsGen`. 
-
-	If automaticAnimation is False (default) then press the SPACE key 
-	to step the animation forward.
-
-	If automaticAnimation is True then the animation will play automatically
-	until all the roots have been found.
-
-	If saveFile is given as a string then the anmation will be saved
-	"""
-	import matplotlib.pyplot as plt
-	from matplotlib import animation
-	fig = plt.gcf()
-	ax = plt.gca()
-
-	rootFinder = findRootsGen(originalContour, f, df, **kwargs)
-	originalContour.plot(linecolor='k', linestyle='--')
-
-	### XXX: show total number of roots to be found at the start
-	# ax.text(0.02, 0.95, 'Zeros remaining: %i'%originalContour.count_roots(f,df,**kwargs), transform=ax.transAxes)
-
-	def update_frame(args):
-		roots, multiplicities, boxes, numberOfRemainingRoots = args
-		# print(args)
-
-		plt.cla() # clear axis
-		originalContour.plot(linecolor='k', linestyle='--')
-		originalContour.sizePlot()
-		for box, numberOfEnclosedRoots in boxes:
-			if not hasattr(box, '_color'):
-				cmap = plt.get_cmap('jet')
-				box._color = cmap(np.random.random())
-			
-			plt.text(box.centralPoint.real, box.centralPoint.imag, numberOfEnclosedRoots)
-			box.plot(linecolor=box._color)
-
-		plt.scatter(np.real(roots), np.imag(roots))
-
-		rootsLabel = ax.text(0.02, 0.95, 'Zeros remaining: %i'%numberOfRemainingRoots, transform=ax.transAxes)
-
-		fig.canvas.draw()
-
-	if saveFile:
-		automaticAnimation = True
-
-	if automaticAnimation:
-		anim = animation.FuncAnimation(fig, update_frame, frames=rootFinder, interval=500)
-
-	else:
-		def draw_next(event):
-			if event.key == ' ':
-				update_frame(next(rootFinder)) 
-		fig.canvas.mpl_connect('key_press_event', draw_next)
-
-	if saveFile:
-		anim.save(filename=saveFile, fps=1, dpi=200)
-		plt.close()
-	else:
-		plt.show()
-
-
-def showRoots(originalContour, f, df=None, **kwargs):
-	"""
-	Plots all roots of a given function f within a given originalContour.  
-	Shares key word arguments with :func:`cxroots.RootFinder.findRootsGen`.
-	"""
-	import matplotlib.pyplot as plt
-	originalContour.plot(linecolor='k', linestyle='--')
-	roots, multiplicities = findRoots(originalContour, f, df, **kwargs)
-	plt.scatter(np.real(roots), np.imag(roots), color='k', marker='x')
-	plt.show()
-
-def printRoots(*args, **kwargs):
-	roots, multiplicities = findRoots(*args, **kwargs)
-
-	# reorder roots
-	sortargs = np.argsort(roots)
-	roots, multiplicities = np.array(roots), np.array(multiplicities)
-	roots, multiplicities = roots[sortargs], multiplicities[sortargs]
-
-	print(' Multiplicity |               Root              ')
-	print('------------------------------------------------')
-	for i, root in enumerate(roots):
-		if root.real < 0:
-			print('{: ^14d}| {:.12f} {:+.12f}i'.format(int(multiplicities[i]), root.real, root.imag))
-		else:
-			print('{: ^14d}|  {:.12f} {:+.12f}i'.format(int(multiplicities[i]), root.real, root.imag))
-
+	return RootResult(roots, multiplicities, originalContour)
