+++ cxroots/__init__.py
@@ -1,4 +1,4 @@
 from .Contours import Circle, Annulus, PolarRect, Rectangle
-from .RootFinder import findRoots, demo_findRoots, showRoots
+from .RootFinder import findRoots, demo_findRoots
 from .CountRoots import count_enclosed_roots
 from .CxDerivative import CxDeriv
